fix(parser): Normalize SMS body to handle multi-line messages
